+++ commons/src/commonTest/kotlin/com/vitorpamplona/amethyst/commons/model/nip05DnsIdentifiers/namecoin/NamecoinSettingsTest.kt
@@ -18,7 +18,7 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.service.namecoin
+package com.vitorpamplona.amethyst.commons.model.nip05DnsIdentifiers.namecoin
 
 import com.vitorpamplona.quartz.nip05DnsIdentifiers.namecoin.ElectrumxServer
 import org.junit.Assert.assertEquals
@@ -116,14 +116,12 @@ class NamecoinSettingsTest {
 
     @Test
     fun `returns parsed list for valid custom servers`() {
-        val settings =
-            NamecoinSettings(
-                customServers =
-                    listOf(
-                        "server1.com:50006",
-                        "server2.onion:50001:tcp",
-                    ),
-            )
+        val settings = NamecoinSettings(
+            customServers = listOf(
+                "server1.com:50006",
+                "server2.onion:50001:tcp",
+            ),
+        )
         val servers = settings.toElectrumxServers()
         assertNotNull(servers)
         assertEquals(2, servers!!.size)
@@ -136,15 +134,13 @@ class NamecoinSettingsTest {
 
     @Test
     fun `skips invalid entries in custom server list`() {
-        val settings =
-            NamecoinSettings(
-                customServers =
-                    listOf(
-                        "valid.com:50006",
-                        "invalid", // no port
-                        "also-invalid:abc", // non-numeric port
-                    ),
-            )
+        val settings = NamecoinSettings(
+            customServers = listOf(
+                "valid.com:50006",
+                "invalid", // no port
+                "also-invalid:abc", // non-numeric port
+            ),
+        )
         val servers = settings.toElectrumxServers()
         assertNotNull(servers)
         assertEquals(1, servers!!.size)
